add set_up_core_dir and move lock_dir_until_exit
