+++ certbot/log.py
@@ -131,7 +131,7 @@ def setup_log_file_handler(config, logfile, fmt):
     """
     # TODO: logs might contain sensitive data such as contents of the
     # private key! #525
-    util.make_or_verify_core_dir(
+    util.set_up_core_dir(
         config.logs_dir, 0o700, os.geteuid(), config.strict_permissions)
     log_file_path = os.path.join(config.logs_dir, logfile)
     try:
